Clean up faction removal logic

+++ src/game/public/game.h
@@ -136,7 +136,8 @@ void   G_SafeFree(struct entity *ent);
 bool   G_AddFaction(const char *name, vec3_t color);
 bool   G_RemoveFaction(int faction_id);
 bool   G_UpdateFaction(int faction_id, const char *name, vec3_t color, bool control);
-int    G_GetFactions(char out_names[][MAX_FAC_NAME_LEN], vec3_t *out_colors, bool *out_ctrl);
+uint16_t G_GetFactions(char out_names[][MAX_FAC_NAME_LEN], vec3_t *out_colors, bool *out_ctrl);
+
 bool   G_SetDiplomacyState(int fac_id_a, int fac_id_b, enum diplomacy_state ds);
 bool   G_GetDiplomacyState(int fac_id_a, int fac_id_b, enum diplomacy_state *out);
